Fixed BoxWebHookTest

+++ src/test/java/com/box/sdk/BoxWebHookTest.java
@@ -9,19 +9,21 @@ import java.util.HashSet;
 import java.util.Iterator;
 import java.util.Set;
 
+import static com.box.sdk.BoxWebHook.Trigger.FOLDER_COPIED;
+import static com.box.sdk.BoxWebHook.Trigger.FOLDER_DOWNLOADED;
+import static com.box.sdk.UniqueTestFolder.*;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasItem;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.notNullValue;
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThat;
 
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit.WireMockClassRule;
 import org.hamcrest.Matchers;
-import org.junit.Assert;
-import org.junit.ClassRule;
-import org.junit.Test;
+import org.junit.*;
 import org.junit.experimental.categories.Category;
 
 import com.eclipsesource.json.JsonArray;
@@ -34,6 +36,16 @@ import javax.swing.*;
  */
 public class BoxWebHookTest {
 
+    @BeforeClass
+    public static void setup() {
+        setupUniqeFolder();
+    }
+
+    @AfterClass
+    public static void afterClass() {
+        removeUniqueFolder();
+    }
+
     @ClassRule
     public static final WireMockClassRule WIRE_MOCK_CLASS_RULE = new WireMockClassRule(53621);
     private BoxAPIConnection api = TestConfig.getAPIConnection();
@@ -42,14 +54,12 @@ public class BoxWebHookTest {
     @Category(IntegrationTest.class)
     public void createWebHookFileSucceeds() throws IOException {
         BoxAPIConnection api = new BoxAPIConnection(TestConfig.getAccessToken());
-        BoxFolder rootFolder = BoxFolder.getRootFolder(api);
         String fileName = "[createWebhook] Test File.txt";
-        byte[] fileBytes = "Non-empty string".getBytes(StandardCharsets.UTF_8);
 
-        InputStream uploadStream = new ByteArrayInputStream(fileBytes);
-        BoxFile uploadedFile = rootFolder.uploadFile(uploadStream, fileName).getResource();
+        BoxFile uploadedFile = null;
 
         try {
+            uploadedFile = uploadFileToUniqueFolderWithSomeContent(api, fileName);
             URL address = new URL("https://www.google.com");
             BoxWebHook.Info info = BoxWebHook.create(uploadedFile, address,
                     BoxWebHook.Trigger.FILE_PREVIEWED, BoxWebHook.Trigger.FILE_LOCKED);
@@ -63,7 +73,9 @@ public class BoxWebHookTest {
 
             info.getResource().delete();
         } finally {
-            uploadedFile.delete();
+            if (uploadedFile != null) {
+                uploadedFile.delete();
+            }
         }
     }
 
@@ -71,25 +83,27 @@ public class BoxWebHookTest {
     @Category(IntegrationTest.class)
     public void createWebHookFolderSucceeds() throws IOException {
         BoxAPIConnection api = new BoxAPIConnection(TestConfig.getAccessToken());
-        BoxFolder rootFolder = BoxFolder.getRootFolder(api);
+        BoxFolder rootFolder = getUniqueFolder(api);
         String folderName = "[createWebhook] Folder";
-        BoxFolder folder = rootFolder.createFolder(folderName).getResource();
+        BoxFolder folder = null;
 
         try {
+            folder = rootFolder.createFolder(folderName).getResource();
             URL address = new URL("https://www.google.com");
-            BoxWebHook.Info info = BoxWebHook.create(folder, address,
-                    BoxWebHook.Trigger.FOLDER_DOWNLOADED, BoxWebHook.Trigger.FOLDER_COPIED);
+            BoxWebHook.Info info = BoxWebHook.create(folder, address, FOLDER_DOWNLOADED, FOLDER_COPIED);
 
             assertThat(info.getID(), is(notNullValue()));
             assertThat(info.getAddress(), is(equalTo(address)));
             assertThat(info.getTarget().getType(), is(equalTo(BoxResource.getResourceType(BoxFolder.class))));
             assertThat(info.getTarget().getId(), is(equalTo(folder.getID())));
-            assertThat(info.getTriggers(), is(equalTo(this.toSet(new BoxWebHook.Trigger[]{
-                BoxWebHook.Trigger.FOLDER_COPIED, BoxWebHook.Trigger.FOLDER_DOWNLOADED}))));
+            assertThat(info.getTriggers(),
+                    is(equalTo(this.toSet(new BoxWebHook.Trigger[]{FOLDER_COPIED, FOLDER_DOWNLOADED}))));
 
             info.getResource().delete();
         } finally {
-            folder.delete(true);
+            if (folder != null) {
+                folder.delete(true);
+            }
         }
     }
 
@@ -97,14 +111,11 @@ public class BoxWebHookTest {
     @Category(IntegrationTest.class)
     public void listWebHooksSucceeds() throws IOException {
         BoxAPIConnection api = new BoxAPIConnection(TestConfig.getAccessToken());
-        BoxFolder rootFolder = BoxFolder.getRootFolder(api);
         String fileName = "[listWebhooks] Test File.txt";
-        byte[] fileBytes = "Non-empty string".getBytes(StandardCharsets.UTF_8);
-
-        InputStream uploadStream = new ByteArrayInputStream(fileBytes);
-        BoxFile uploadedFile = rootFolder.uploadFile(uploadStream, fileName).getResource();
+        BoxFile uploadedFile = null;
 
         try {
+            uploadedFile = uploadFileToUniqueFolderWithSomeContent(api, fileName);
             URL address = new URL("https://www.google.com");
             BoxWebHook.Info info = BoxWebHook.create(uploadedFile, address, BoxWebHook.Trigger.FILE_PREVIEWED);
             Iterable<BoxWebHook.Info> webhooks = BoxWebHook.all(api);
@@ -113,10 +124,12 @@ public class BoxWebHookTest {
 
             info.getResource().delete();
 
-            webhooks = BoxWebHook.all(this.api);
+            webhooks = BoxWebHook.all(api);
             assertThat(webhooks, not(hasItem(Matchers.<BoxWebHook.Info>hasProperty("ID", equalTo(info.getID())))));
         } finally {
-            uploadedFile.delete();
+            if (uploadedFile != null) {
+                uploadedFile.delete();
+            }
         }
     }
 
@@ -158,13 +171,11 @@ public class BoxWebHookTest {
     @Test
     @Category(UnitTest.class)
     public void testCreateWebhookSucceedsAndSendsCorrectJson() throws IOException {
-        String result = "";
         final String webhookID = "12345";
         final String folderID = "1111";
         final String createdByID = "2222";
         final String createdByLogin = "test@user.com";
         final String address = "https:/example.com";
-        final String trigger = "FILE_LOCKED";
         final String webhookURL = "/webhooks";
 
         JsonObject innerObject = new JsonObject()
@@ -176,7 +187,7 @@ public class BoxWebHookTest {
                 .add("triggers", new JsonArray().add("FILE.LOCKED"))
                 .add("address", address);
 
-        result = TestConfig.getFixture("BoxWebhook/CreateWebhook201");
+        String result = TestConfig.getFixture("BoxWebhook/CreateWebhook201");
 
         WIRE_MOCK_CLASS_RULE.stubFor(WireMock.post(WireMock.urlPathEqualTo(webhookURL))
                 .withRequestBody(WireMock.equalToJson(webhookObject.toString()))
@@ -187,24 +198,22 @@ public class BoxWebHookTest {
         BoxFolder folder = new BoxFolder(this.api, folderID);
         BoxWebHook.Info webhookInfo = BoxWebHook.create(folder, new URL(address), BoxWebHook.Trigger.FILE_LOCKED);
 
-        Assert.assertEquals(webhookID, webhookInfo.getID());
-        Assert.assertEquals(folderID, webhookInfo.getTarget().getId());
-        Assert.assertEquals(createdByID, webhookInfo.getCreatedBy().getID());
-        Assert.assertEquals(createdByLogin, webhookInfo.getCreatedBy().getLogin());
-        Assert.assertEquals(BoxWebHook.Trigger.FILE_LOCKED, webhookInfo.getTriggers().iterator().next());
+        assertEquals(webhookID, webhookInfo.getID());
+        assertEquals(folderID, webhookInfo.getTarget().getId());
+        assertEquals(createdByID, webhookInfo.getCreatedBy().getID());
+        assertEquals(createdByLogin, webhookInfo.getCreatedBy().getLogin());
+        assertEquals(BoxWebHook.Trigger.FILE_LOCKED, webhookInfo.getTriggers().iterator().next());
     }
 
     @Test
     @Category(UnitTest.class)
     public void testGetWebhookSuceeds() throws IOException {
-        String result = "";
         final String webhookID = "12345";
         final String folderID = "1111";
         final String createdByID = "2222";
-        final String address = "https://example.com";
         final String webhookURL = "/webhooks/" + webhookID;
 
-        result = TestConfig.getFixture("BoxWebhook/GetWebhook200");
+        String result = TestConfig.getFixture("BoxWebhook/GetWebhook200");
 
         WIRE_MOCK_CLASS_RULE.stubFor(WireMock.get(WireMock.urlPathEqualTo(webhookURL))
                 .willReturn(WireMock.aResponse()
@@ -214,20 +223,19 @@ public class BoxWebHookTest {
         BoxWebHook webhook = new BoxWebHook(this.api, webhookID);
         BoxWebHook.Info info = webhook.getInfo();
 
-        Assert.assertEquals(webhookID, info.getID());
-        Assert.assertEquals(folderID, info.getTarget().getId());
-        Assert.assertEquals(createdByID, info.getCreatedBy().getID());
+        assertEquals(webhookID, info.getID());
+        assertEquals(folderID, info.getTarget().getId());
+        assertEquals(createdByID, info.getCreatedBy().getID());
     }
 
     @Test
     @Category(UnitTest.class)
     public void testGetAllWebhooksSucceeds() throws IOException {
-        String result = "";
         final String webhookID = "12345";
         final String folderID = "1111";
         final String webhookURL = "/webhooks";
 
-        result = TestConfig.getFixture("BoxWebhook/GetAllWebhooks200");
+        String result = TestConfig.getFixture("BoxWebhook/GetAllWebhooks200");
 
         WIRE_MOCK_CLASS_RULE.stubFor(WireMock.get(WireMock.urlPathEqualTo(webhookURL))
                 .willReturn(WireMock.aResponse()
@@ -237,30 +245,26 @@ public class BoxWebHookTest {
         Iterator<BoxWebHook.Info> webhooks = BoxWebHook.all(this.api).iterator();
         BoxWebHook.Info webhookInfo = webhooks.next();
 
-        Assert.assertEquals(webhookID, webhookInfo.getID());
-        Assert.assertEquals(folderID, webhookInfo.getTarget().getId());
+        assertEquals(webhookID, webhookInfo.getID());
+        assertEquals(folderID, webhookInfo.getTarget().getId());
     }
 
     @Test
     @Category(UnitTest.class)
     public void testUpdateWebhookInfoSucceeds() throws IOException {
-        String result = "";
-        String getResult = "";
         final String newAddress = "https://newexample.com";
         final String webhookID = "12345";
         final String createdByLogin = "test@user.com";
         final String webhookURL = "/webhooks/" + webhookID;
 
         JsonArray triggers = new JsonArray()
-            .add("FILE.UPLOADED");
+                .add("FILE.UPLOADED");
 
         JsonObject updateObject = new JsonObject()
                 .add("triggers", triggers)
                 .add("address", newAddress);
 
-        getResult = TestConfig.getFixture("BoxWebhook/GetWebhook200");
-
-        result = TestConfig.getFixture("BoxWebhook/UpdateWebhookInfo200");
+        String result = TestConfig.getFixture("BoxWebhook/UpdateWebhookInfo200");
 
         WIRE_MOCK_CLASS_RULE.stubFor(WireMock.get(WireMock.urlPathEqualTo(webhookURL))
                 .willReturn(WireMock.aResponse()
@@ -279,28 +283,27 @@ public class BoxWebHookTest {
         info.setTriggers(BoxWebHook.Trigger.FILE_UPLOADED);
         webhook.updateInfo(info);
 
-        Assert.assertEquals(new URL(newAddress), info.getAddress());
-        Assert.assertEquals(webhookID, info.getID());
-        Assert.assertEquals(createdByLogin, info.getCreatedBy().getLogin());
+        assertEquals(new URL(newAddress), info.getAddress());
+        assertEquals(webhookID, info.getID());
+        assertEquals(createdByLogin, info.getCreatedBy().getLogin());
     }
 
     @Test
     @Category(UnitTest.class)
-    public void testDeleteWebhookSucceeds() throws IOException {
-        String result = "";
+    public void testDeleteWebhookSucceeds() {
         final String webhookID = "12345";
         final String webhookURL = "/webhooks/" + webhookID;
 
         WIRE_MOCK_CLASS_RULE.stubFor(WireMock.delete(WireMock.urlPathEqualTo(webhookURL))
                 .willReturn(WireMock.aResponse()
                         .withHeader("Content-Type", "application/json")
-                        .withBody(result)));
+                        .withBody("")));
 
         BoxWebHook webhook = new BoxWebHook(this.api, webhookID);
         webhook.delete();
     }
 
     private <T> Set<T> toSet(T[] values) {
-        return new HashSet<T>(Arrays.asList(values));
+        return new HashSet<>(Arrays.asList(values));
     }
 }
